accept partial query state
